requirements.txt
Flask==3.1.2
gunicorn==23.0.0
requests==2.32.5
python-dotenv==1.1.1
Werkzeug==3.1.5
